Patch #2018 by Rick: * Added CR to clipboard when exporting compiler messages on Win32 * Removed signed/unsigned warnings from various dialogs
git-svn-id: https://svn.code.sf.net/p/codeblocks/code/trunk@3992 2a5c6006-c6dd-42ca-98ab-0921f2732cef

+++ src/sdk/configuretoolsdlg.cpp
@@ -93,7 +93,7 @@ void ConfigureToolsDlg::OnUpdateUI(wxUpdateUIEvent& /*event*/)
 	const wxListBox* list = XRCCTRL(*this, "lstTools", wxListBox);
 	bool hasSel = list->GetSelection() != -1;
 	bool notFirst = list->GetSelection() > 0;
-	bool notLast = (list->GetSelection() < list->GetCount() -1) && hasSel;
+	bool notLast = (list->GetSelection() < (int)(list->GetCount()) -1) && hasSel;
 	bool notSeparator = true;
 
 	if(hasSel)
